+++ src/test/java/test/org/fugerit/java/nhg/TestGenerateLibraryConfiguration.java
@@ -1,16 +1,12 @@
 package test.org.fugerit.java.nhg;
 
 import lombok.extern.slf4j.Slf4j;
-import org.fugerit.java.nhg.GenerateReflectConfig;
 import org.fugerit.java.nhg.config.NativeHelperFacade;
 import org.fugerit.java.nhg.config.model.NativeHelperConfig;
-import org.fugerit.java.nhg.reflect.config.Entry;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.io.StringWriter;
-import java.util.List;
 
 @Slf4j
 class TestGenerateLibraryConfiguration {